Revert "cogl-gst: add missing Cogl introspection dependency"
The pkg-config file might not necessarily be installed yet so I don't
think it makes sense to try and include it when running g-ir-scanner.
Presumably it should pick up the headers from source directory instead
of the installed directory. It seems to build without this patch so
let's just revert it.

This reverts commit d9c8570f14a43e6c3a48c521e819427f02e816e4.